(wip) extract filter for ECJ bytecode into separate class
