+++ org.jacoco.core/src/org/jacoco/core/internal/analysis/MethodAnalyzer.java
@@ -22,6 +22,7 @@ import org.jacoco.core.analysis.ISourceNode;
 import org.jacoco.core.internal.analysis.filter.IFilter;
 import org.jacoco.core.internal.analysis.filter.IFilterOutput;
 import org.jacoco.core.internal.analysis.filter.SynchronizedFilter;
+import org.jacoco.core.internal.analysis.filter.TryWithResourcesEcjFilter;
 import org.jacoco.core.internal.analysis.filter.TryWithResourcesFilter;
 import org.jacoco.core.internal.flow.IFrame;
 import org.jacoco.core.internal.flow.Instruction;
@@ -42,7 +43,8 @@ public class MethodAnalyzer extends MethodProbesVisitor
 		implements IFilterOutput {
 
 	private static final IFilter[] FILTERS = new IFilter[] {
-			new SynchronizedFilter(), new TryWithResourcesFilter() };
+			new SynchronizedFilter(), new TryWithResourcesFilter(),
+			new TryWithResourcesEcjFilter() };
 
 	private final boolean[] probes;
